[FLINK-1159] Moved helpers to different packages
